services/climate_control/requirements.txt
tapo
paho-mqtt
pyyaml
